rename: input parameter files
