--- src/python/impactx/dashboard/Input/inputParameters/__init__.py
@@ -1,5 +1,0 @@
-from .inputFunctions import InputFunctions
-
-__all__ = [
-    "InputFunctions",
-]
