+++ src/python/impactx/dashboard/Input/input_parameters/__init__.py
@@ -1,0 +1,5 @@
+from .utils import InputFunctions
+
+__all__ = [
+    "InputFunctions",
+]
